gta-warehouse-demo/scripts/requirements.txt
pandas==2.0.3
numpy==1.24.3
faker==19.12.0
psycopg2-binary==2.9.9
scikit-learn==1.3.2
xgboost==2.0.2
prophet==1.1.5
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0